refactor(tests): split the 6,201-line booking test file by module
`services/booking/tests.rs` held 156 functions in one file while the code
it tests has been split into eight modules for months, and its own
fixtures already live in `tests/support/`. A failure named `tests::` and
nothing more.

Split into `tests/mod.rs` plus 14 modules mirroring the production
layout, largest first: reservation_idempotency 19 tests, reservations 16,
reporting 16, group_idempotency 16, stays 13, folio 13, payments 12,
checkout_settlement 11, groups 9, pricing 6, money_guards 5,
group_services 5, extend_stay 5, guests 3.

Shared imports live in a `prelude` glob-imported by each submodule. A
glob import never reports unused imports, so `-D warnings` stays
satisfied without maintaining 14 separate import lists by hand.

The move is mechanical. Verified rather than assumed: line-range
coverage was asserted programmatically so every line lands in exactly one
module with no gaps or overlaps, and every function body was diffed
against the previous revision — 156 before, 156 after, none added or
removed, all bodies byte-identical except the one change below. The seven
local fixtures were placed by their actual usage ranges, which is why the
stays/checkout_settlement boundary sits at the `checkout_dt` definition
rather than at the first settlement test.

One deliberate edit: record_payment_returning_id_tx_returns_inserted_-
transaction_id called `super::support::begin_tx`. Two modules are named
`support` — production `services/booking/support.rs` and the test
fixtures — and `super::support` meant the production one from the old
`tests.rs` but resolves to the fixtures from inside `tests::payments`.
Repointed to the absolute path; the compiler caught it.

cargo test 845 passed, 0 failed (149 booking tests, unchanged); clippy
clean; npm test 358 passed; `npm run verify:full` exit 0 with the native
app smoke boot.
